requirements.txt
pandas~=2.2.3
pillow==11.1.0
torch==2.5.1+cu124
transformers==4.48.0
datasets==3.2.0
scikit-learn==1.6.1

langchain-core~=0.3.29
langchain-openai~=0.3.0
python-dotenv~=1.0.1
tqdm~=4.67.1